+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+- Added lineCount and hasRefSeq functions to BamFile, each accepting a string seqName
+- Fixed aborting on index retrieval code
+
+
 ## [1.0.18](https://github.com/GMOD/bam-js/compare/v1.0.17...v1.0.18) (2019-05-01)